Sprint 8: res.json

+++ src/controllers/productosController.js
@@ -117,32 +117,57 @@ let controller = {
 
         let idProducto = req.params.id;
 
-        let nombreImagen = req.file.filename;
+        let errors = validationResult(req);
+        if(errors.errors.length == 0){
 
-        let newDataProduct = {
-            nombre: req.body.nombre,
-            id_medioFk: req.body.medio,
-            id_categoriaFk: req.body.categoria,
-            descripcion: req.body.descripcion,
-            precio: req.body.precio,
-            descuento: req.body.descuento,
-            ancho: req.body.ancho,
-            alto: req.body.alto,
-            otros_detalles: req.body.otros_detalles,
-            anio_creacion: req.body.year,
-            img: nombreImagen
-        };
+            let nombreImagen = req.file.filename;
 
-        db.Producto.findByPk(idProducto)
-            .then(function(producto){
-                let imagenAnterior = producto.img;
-                fs.unlinkSync(path.join(__dirname, '../../public/img', imagenAnterior));  //borro imagen vieja
-            })
-            .then(db.Producto.update(newDataProduct, {where: {id: idProducto}}))
-            .then(function(resultado){
-            res.redirect('/galeria');
-            });
-        },
+            let newDataProduct = {
+                nombre: req.body.nombre,
+                id_medioFk: req.body.medio,
+                id_categoriaFk: req.body.categoria,
+                descripcion: req.body.descripcion,
+                precio: req.body.precio,
+                descuento: req.body.descuento,
+                ancho: req.body.ancho,
+                alto: req.body.alto,
+                otros_detalles: req.body.otros_detalles,
+                anio_creacion: req.body.year,
+                img: nombreImagen
+            };
+
+            db.Producto.findByPk(idProducto)
+                .then(function(producto){
+                    let imagenAnterior = producto.img;
+                    fs.unlinkSync(path.join(__dirname, '../../public/img', imagenAnterior));  //borro imagen vieja
+                })
+                .then(db.Producto.update(newDataProduct, {where: {id: idProducto}}))
+                .then(function(resultado){
+                res.redirect('/galeria');
+                });
+            }else{
+                let allCategorias = db.Categoria.findAll()
+                .then(function (categorias) {
+                    return categorias
+                });
+                let allMedios = db.Medio.findAll()
+                    .then(function (medios) {
+                    return medios
+                });
+        
+                let id = req.params.id;
+        
+                let obra = db.Producto.findByPk(id).then(function(obraData){
+                    return obraData;
+                })
+        
+                Promise.all([allCategorias, allMedios, obra])
+                    .then(function([dataCategorias, dataMedios, dataObra]){
+                        res.render('editar_producto',{categorias: dataCategorias, medios: dataMedios, obraAEditar: dataObra, errors: errors.mapped(), old: req.body });
+                    })
+            }
+        
+    },
 
     delete: (req, res) => {
